pyproject.toml
[project]
name = "ivf-batch"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.13.2",
    "numpy>=2.4.4",
]
